requirements.txt
django==4.2.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.267
langchain-community
pydantic==1.10.8
huggingface-hub==0.16.4
transformers==4.32.0
llama-cpp-python==0.1.77
python-dotenv==1.0.0
tqdm==4.66.1
google-generativeai==0.3.1
python-dotenv==1.0.0
aiohttp==3.9.1
